Reorganized files and added pyproject.toml.
