+++ .gitignore
@@ -3,3 +3,8 @@ __pycache__
 *.png
 *.npy
 *.cube
+
+# Distribution / packaging
+build/
+dist/
+*.egg-info/
