+++ data/molecule_29766_0/GEOM-B3LYP.xyz
