+++ qmtools/__init__.py
@@ -1,0 +1,4 @@
+
+# from .qmtools import *
+
+__version__ = "0.0.1"
