+++ qmtools/automaton_nn_simple.py
