+++ qmtools/cidhelper.py
