+++ qmtools/cuda/kernel_common.h
