+++ qmtools/cuda/kernel_nn_ss.cu
