+++ scripts/test-hartree-fft.py
