+++ scripts/test-hartree-molecule.py
