+++ scripts/test-hartree-point.py
