+++ scripts/testnn.py
